Use empty header comments (#14)
Xcode by default creates a header comment that includes the file's name, the target name, the user who created the file and the date of creation. All of which can be easily located via Git history.

Adding the template macro will unfortunately not get rid of the header comment completely, but it will generate a header comment that does not contain all the boilerplate. Removing the header comment should be part of out tooling (e.g. Swiftformat)